Tidy up multipart
